Restore FP32 comparison precision in compute_mi_exponential
The previous commit replaced torch.exp(float32_tensor) with
math.exp(float64) — numerically more precise, but the optimizer was
calibrated against the original FP32 trajectory: torch.tensor(-mi_val)
defaults to float32 and torch.exp on a float32 tensor produces float32,
so goldsearch's anti-drift guard and best-metric comparisons all ran
in FP32. With FP64 outputs the optimizer can distinguish values that
differ below the float32 ULP and ends up on a slightly different
trajectory — observed RMSE went from ~1.15 to ~1.95 px.

Cast math.exp(...) back to float32 (still stored as a Python float so
the loop arithmetic stays off-torch). Keeps every speed win, restores
the exact comparison precision of the original path.

+++ ETNA/registrators_pyramidal.py
@@ -251,10 +251,15 @@ class EtnaMultiMetric(object):
         mi = self.compute_mi(ref_img, flt_img, t_mat, eref)
         _t = time.perf_counter()
         # math.exp on a Python float — was torch.exp(0-d).cpu() (~5 ms/call).
-        # Monotone, FP64 bit-equivalent.
+        # The original path stored values in float32 (torch.tensor default
+        # dtype + torch.exp on a float32 tensor), so the optimizer compared
+        # FP32-precision metrics. Cast back to float32 here to keep the same
+        # comparison precision: in pure-FP64 the optimizer can distinguish
+        # values that differ below the float32 ULP and ends up on a slightly
+        # different (worse) trajectory.
         if isinstance(mi, torch.Tensor):
             mi = float(mi.item())
-        result = math.exp(mi)
+        result = float(np.float32(math.exp(mi)))
         self._wrap_times["compute_mi_exp"] += time.perf_counter() - _t
         self._wrap_counts["compute_mi_exp"] += 1
         return result
